test: passing test with mocks
